AGeoPoint, AFlatGeoPoint, RoutePlanner: use RoughAltitude
Wrap the compressed altitude in a nice OO interface.

+++ src/Engine/Task/Tasks/AbortTask.cpp
@@ -157,7 +157,7 @@ AbortTask::fill_reachable(const AircraftState &state,
   if (task_full() || approx_waypoints.empty())
     return false;
 
-  const AGeoPoint p_start (state.location, (short)state.altitude);
+  const AGeoPoint p_start(state.location, state.altitude);
 
   bool found_final_glide = false;
   reservable_priority_queue<Alternate, AlternateVector, AbortRank> q;
